pyproject.toml
[project]
name = "free-lunch-ai"
version = "0.1.0"
description = "A plug-and-play router that unifies free LLM APIs. Define models in YAML & explore ideas instantly. Perfect for building agents on a budget"
keywords = ["llm", "agent", "langchain", "groq", "gemini", "free-tier", "router"]
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "tctsung", email = "tctsung@nyu.edu" }
]
license = { text = "MIT" }

# Core: light router only (no LangChain)
dependencies = [
    "httpx>=0.24.0",
    "python-dotenv",
    "PyYAML",
    "pydantic>=2.0.0" 
]

[project.optional-dependencies]
langchain = [
    "langchain-core>=1.0.0,<2.0.0",
    "langchain-groq>=1.0.0,<2.0.0",         
    "langchain-google-genai>=4.0.0,<5.0.0",
    "langchain-openai>=1.0.0,<2.0.0",
]
all = ["free-lunch-ai[langchain]"]

[project.urls]
Homepage = "https://github.com/tctsung/free-lunch-ai"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/free_lunch"]
